feat: added more types
