+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
 	"compilerOptions": {
+		"noEmit": true,
 		"allowJs": true,
 		"checkJs": true
 	}
